divide conftest appropriately
