+++ tests/dags/conftest.py
@@ -1,10 +1,9 @@
 from pathlib import Path
-from unittest.mock import patch
 
 import pytest
 from airflow.models import DagBag
 
-PROJECT_DIR = Path(__file__).parent.parent
+PROJECT_DIR = Path(__file__).parent.parent.parent
 
 
 @pytest.fixture()
@@ -21,17 +20,3 @@ def dagbag() -> DagBag:
         dag_folder=Path(PROJECT_DIR) / "dags",
         include_examples=False,
     )
-
-
-@pytest.fixture
-def mock_ucam_op():
-    """Return a mocked UcamHook imported via ideafast_etl.operators"""
-    with patch("ideafast_etl.operators.ucam.UcamHook") as ucam_hook_mock:
-        yield ucam_hook_mock
-
-
-@pytest.fixture
-def mock_mongo_op():
-    """Return a mocked MongoHook imported via ideafast_etl.operators"""
-    with patch("ideafast_etl.operators.ucam.LocalMongoHook") as mongo_hook_mock:
-        yield mongo_hook_mock
